tornado: move mTLS SSLContext generation into separate module
The code is shared between registrar and verifier.
